build(android): fix android.bundle not included in the release build (#96)
## Description

Closes: #XXXX



This PR updates the build configurations for Android so that the generated `index.android.bundle` is correctly included in the release build, this issue was caused from a bad update from `react-native` 0.66.4 to 0.70.6.  
To fix this problem I have used [react-native-upgrade-helper](https://react-native-community.github.io/upgrade-helper/?from=0.66.4&to=0.70.6) that provides the changes that should be applayed when updating the `react-native` package.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [x] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed