fix: fix splash screen not been visualized on Android after unlocking the device (#114)
## Description

Closes: #XXXX



This PR fixes a bug that cause the missing visualization of the splash screen on Android when the application goes on `blur`. This is caused from the fact that if the application requires the biometrics authentication shortly after the user have unlocked the device the biometrics popup is not showed to the user and so the application don't goes on `blur` state and the `noSplashScreen` flag is not cleared causing the application to not show the splash screen if the user opens another application.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

- [ ] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] provided a link to the relevant issue or specification
- [ ] reviewed "Files changed" and left comments if necessary
- [ ] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed