feat: improve Legder account generation speed (#147)
## Description

Closes: #XXXX

This PR improves the loading speed when importing an account with the Ledger device.To achive this I have reduced the amount of accounts generated for each page from 15 to 10.
This PR fixes also a bug that couse a communication error with the Ledger if the user toggle the `HdPath` picker while the app is generating a page of accounts.



---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed