fix: balance not displayed when returning to the home (#184)
## Description

Closes: #XXXX



This PR fixes a bug where the balances are not displayed when returning to the home.
Ti bug was caused from the fact that the server was returning an error related to the query of data in a block that has yet to be produced.
To fix this I leveregd the apollo cache so that the application now gently fallback to the last fetched data, in this case the app will just show a non up to date information instead of nothing, the information can be later updated from the user with the pull to refresh gesture. This also have the benefit of making the UI showing the content loaders since the `loading` flag of apollo will correctly refelect the status of the request unlike before.
This is the final result:
![screen-20230526-164230 mp4](https://github.com/desmos-labs/dpm/assets/6245917/75b83fcf-e389-45cf-9a74-7ed23530b1bf)

---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed