feat: improve transaction fee estimation functionality (#193)
## Description

Closes: DPM-73, DPM-72

This PR improves the fee estimation procedure by enforcing a timeout
that if exceed the application will fallback to a default fee value.
The default fee amount and allowed execution time can be configured with
some feature flags from Posthog.
On top of this I have added an analytics to track the amount of time
that the application is abble to correctly estimate the fee in the
provided time to be able to tune the timeout value. This analytics can
be enabled/disbale through a Posthog feature flag.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is
not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable
and please add
your handle next to the items reviewed if you only reviewed selected
items.*

- [ ] confirmed the correct [type
prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json)
in the PR title
- [ ] confirmed all author checklist items have been addressed

---------

Co-authored-by: Riccardo <riccardo.montagnin@gmail.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>