fix: profile refresh when returning to the Profile screen (#215)
## Description

Closes: DPM-122



This PR fixes the profile refresh when returning to the Profile screen. 
To fix this I have removed the `useFocusEffect` that was triggering the refresh and update the `useStoreProfile` hook
to update the user's cached profile after the transaction succed to avoid to refetch the user's profile when returning in the Profile screen after the transaction.

---

### Author Checklist

*All items are required. Please add a note to the item if the item is not applicable and
please add links to any relevant follow up issues.*

- [x] included the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] provided a link to the relevant issue or specification
- [x] reviewed "Files changed" and left comments if necessary
- [x] confirmed all CI checks have passed

### Reviewers Checklist

*All items are required. Please add a note if the item is not applicable and please add
your handle next to the items reviewed if you only reviewed selected items.*

- [ ] confirmed the correct [type prefix](https://github.com/commitizen/conventional-commit-types/blob/v3.0.0/index.json) in the PR title
- [ ] confirmed all author checklist items have been addressed